+++ src/helpers/commonTypes.ts
@@ -1,0 +1,11 @@
+export enum CommonLanguages {
+    // Pick<SupportedLocales, 'plPL' | 'enUS'>
+    /**
+     * en translation
+     */
+    plPL = 'plPL',
+    /**
+     * pl translation
+     */
+    enUS = 'enUS'
+}