+++ src/theme/config/breakpoints.ts
@@ -1,0 +1,7 @@
+export const breakpoints = {
+    xl: 1536,
+    lg: 1200,
+    md: 900,
+    sm: 600,
+    xs: 444
+};