+++ src/theme/config/palette.ts
@@ -1,0 +1,443 @@
+export const palette = {
+    light: {
+        mode: 'light',
+        text: {
+            primary: 'rgba(38, 38, 38, 1)',
+            secondary: 'rgba(123, 123, 123, 1)',
+            disabled: 'rgba(157, 157, 157, 1)',
+            hover: 'rgba(245, 245, 245, 1)',
+            selected: 'rgba(233, 233, 233, 1)',
+            focus: 'rgba(217, 217, 217, 1)',
+            focusVisible: 'rgba(157, 157, 157, 1)'
+        },
+        primary: {
+            main: 'rgba(0, 167, 238, 1)',
+            dark: 'rgba(0, 134, 204, 1)',
+            light: 'rgba(0, 180, 240, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(224, 245, 253, 1)',
+            focus: 'rgba(176, 228, 249, 1)',
+            focusVisible: 'rgba(224, 245, 253, 1)',
+            outlinedBorder: 'rgba(0, 167, 238, 1)',
+            contrastText: 'rgba(255, 255, 255, 1)'
+        },
+        secondary: {
+            main: 'rgba(85, 85, 85, 1)',
+            dark: 'rgba(0, 0, 0, 1)',
+            light: 'rgba(85, 85, 85, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            hover: 'rgba(245, 245, 245, 1)',
+            selected: 'rgba(233, 233, 233, 1)',
+            focus: 'rgba(217, 217, 217, 1)',
+            focusVisible: 'rgba(196, 196, 196, 1)',
+            outlinedBorder: 'rgba(38, 38, 38, 1)'
+        },
+        action: {
+            active: 'rgba(123, 123, 123, 1)',
+            hover: 'rgba(0, 0, 0, 0.1)',
+            selected: 'rgba(233, 233, 233, 1)',
+            disabledBackground: 'rgba(233, 233, 233, 1)',
+            focus: 'rgba(233, 233, 233, 1)',
+            disabled: 'rgba(196, 196, 196, 1)'
+        },
+        error: {
+            main: 'rgba(235, 30, 50, 1)',
+            dark: 'rgba(188, 24, 40, 1)',
+            light: 'rgba(239, 75, 91, 1)',
+            hover: 'rgba(253, 233, 235, 1)',
+            selected: 'rgba(251, 210, 214, 1)',
+            focusVisible: 'rgba(243, 120, 132, 1)',
+            outlinedBorder: 'rgba(235, 30, 50, 1)',
+            contrast: 'rgba(255, 255, 255, 1)'
+        },
+        warning: {
+            main: 'rgba(255, 210, 30, 1)',
+            dark: 'rgba(204, 168, 24, 1)',
+            light: 'rgba(255, 219, 75, 1)',
+            hover: 'rgba(255, 251, 233, 1)',
+            selected: 'rgba(255, 246, 210, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            focusVisible: 'rgba(255, 228, 120, 1)',
+            outlinedBorder: 'rgba(255, 210, 30, 1)'
+        },
+        info: {
+            main: 'rgba(0, 167, 238, 1)',
+            dark: 'rgba(0, 153, 223, 1)',
+            light: 'rgba(0, 180, 240, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(176, 228, 249, 1)',
+            focusVisible: 'rgba(224, 245, 253, 1)',
+            outlinedBorder: 'rgba(0, 167, 238, 1)',
+            contrast: 'rgba(255, 255, 255, 1)'
+        },
+        success: {
+            main: 'rgba(0, 160, 60, 1)',
+            dark: 'rgba(0, 128, 48, 1)',
+            light: 'rgba(51, 179, 99, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            hover: 'rgba(230, 246, 236, 1)',
+            selected: 'rgba(204, 236, 216, 1)',
+            focusVisible: 'rgba(102, 198, 138, 1)',
+            outlinedBorder: 'rgba(0, 160, 60, 1)'
+        },
+        common: {
+            white: 'rgba(255, 255, 255, 1)',
+            black: 'rgba(0, 0, 0, 1)',
+            grey: 'rgba(233, 233, 233, 1)'
+        },
+        background: {
+            'default': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-0': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-1': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-2': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-3': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-4': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-5': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-6': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-7': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-8': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-9': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-10': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-11': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-12': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-13': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-14': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-15': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-16': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-17': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-18': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-19': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-20': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-21': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-22': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-23': 'rgba(255, 255, 255, 1)',
+            'paper-elevation-24': 'rgba(255, 255, 255, 1)'
+        },
+        components: {
+            rating: {
+                enabledBorder: 'rgba(217, 217, 217, 1)',
+                activeFill: 'rgba(255, 210, 30, 1)'
+            },
+            avatar: {
+                fill: 'rgba(157, 157, 157, 1)'
+            },
+            input: {
+                standard: {
+                    enabledBorder: 'rgba(217, 217, 217, 1)',
+                    hoverBorder: 'rgba(0, 0, 0, 1)'
+                },
+                filled: {
+                    enabledFill: 'rgba(245, 245, 245, 1)',
+                    hoverFill: 'rgba(245, 245, 245, 1)'
+                },
+                outlined: {
+                    enabledBorder: 'rgba(217, 217, 217, 1)',
+                    hoverBorder: 'rgba(0, 0, 0, 1)'
+                }
+            },
+            switch: {
+                knobFillEnabled: 'rgba(123, 123, 123, 1)',
+                slideFill: 'rgba(217, 217, 217, 1)',
+                knowFillDisabled: 'rgba(217, 217, 217, 1)'
+            },
+            chip: {
+                defaultCloseFill: 'rgba(0, 0, 0, 1)',
+                defaultHoverFill: 'rgba(245, 245, 245, 1)',
+                defaultEnabledBorder: 'rgba(217, 217, 217, 1)',
+                defaultFocusFill: 'rgba(233, 233, 233, 1)'
+            },
+            tooltip: {
+                fill: 'rgba(0, 0, 0, 1)'
+            },
+            backdrop: {
+                fill: 'rgba(0, 0, 0, 0.5)'
+            },
+            appBar: {
+                defaultFill: 'rgba(233, 233, 233, 1)'
+            },
+            breadcrumbs: {
+                collapseFill: 'rgba(233, 233, 233, 1)'
+            },
+            alert: {
+                error: {
+                    color: 'rgba(38, 38, 38, 1)',
+                    background: 'rgba(253, 233, 235, 1)'
+                },
+                warning: {
+                    color: 'rgba(38, 38, 38, 1)',
+                    background: 'rgba(255, 251, 233, 1)'
+                },
+                info: {
+                    color: 'rgba(38, 38, 38, 1)',
+                    background: 'rgba(224, 245, 253, 1)'
+                },
+                success: {
+                    color: 'rgba(38, 38, 38, 1)',
+                    background: 'rgba(230, 246, 236, 1)'
+                }
+            },
+            stepper: {
+                connector: 'rgba(157, 157, 157, 1)'
+            },
+            snackbar: {
+                fill: 'rgba(38, 38, 38, 1)'
+            },
+            mediaImage: {
+                actions: 'rgba(196, 196, 196, 1)',
+                focusable: 'rgba(0, 0, 0, 0.1)'
+            },
+            datagrid: {
+                border: 'rgba(217, 217, 217, 1)'
+            },
+            skeleton: {
+                border: 'rgba(217, 217, 217, 1)'
+            },
+            detail: {
+                close: {
+                    background: 'rgba(255, 255, 255, 1)',
+                    hover: 'rgba(233, 233, 233, 1)'
+                }
+            }
+        },
+        elevation: {
+            outlined: 'rgba(217, 217, 217, 1)',
+            Color: 'rgba(255, 255, 255, 1)'
+        },
+        _native: {
+            'scrollbar-bg': 'rgba(233, 233, 233, 1)'
+        },
+        contrast: {
+            contrastText: 'rgba(0, 167, 238, 1)',
+            main: 'rgba(255, 255, 255, 1)',
+            dark: 'rgba(255, 255, 255, 1)',
+            light: 'rgba(255, 255, 255, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(224, 245, 253, 1)',
+            focus: 'rgba(176, 228, 249, 1)',
+            focusVisible: 'rgba(224, 245, 253, 1)',
+            outlinedBorder: 'rgba(224, 245, 253, 1)'
+        }
+    },
+    dark: {
+        mode: 'dark',
+        text: {
+            primary: 'rgba(233, 233, 233, 1)',
+            secondary: 'rgba(217, 217, 217, 1)',
+            disabled: 'rgba(157, 157, 157, 1)',
+            hover: 'rgba(217, 217, 217, 1)',
+            selected: 'rgba(196, 196, 196, 1)',
+            focus: 'rgba(196, 196, 196, 1)',
+            focusVisible: 'rgba(157, 157, 157, 1)'
+        },
+        primary: {
+            main: 'rgba(0, 167, 238, 1)',
+            dark: 'rgba(0, 153, 223, 1)',
+            light: 'rgba(0, 180, 240, 1)',
+            contrast: 'rgba(196, 196, 196, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(224, 245, 253, 1)',
+            focus: 'rgba(0, 117, 184, 1)',
+            focusVisible: 'rgba(0, 117, 184, 1)',
+            outlinedBorder: 'rgba(0, 167, 238, 1)',
+            contrastText: 'rgba(38, 38, 38, 1)'
+        },
+        secondary: {
+            main: 'rgba(217, 217, 217, 1)',
+            dark: 'rgba(157, 157, 157, 1)',
+            light: 'rgba(245, 245, 245, 1)',
+            contrast: 'rgba(38, 38, 38, 1)',
+            hover: 'rgba(245, 245, 245, 1)',
+            selected: 'rgba(233, 233, 233, 1)',
+            focus: 'rgba(217, 217, 217, 1)',
+            focusVisible: 'rgba(196, 196, 196, 1)',
+            outlinedBorder: 'rgba(38, 38, 38, 1)'
+        },
+        action: {
+            active: 'rgba(245, 245, 245, 1)',
+            hover: 'rgba(67, 67, 67, 1)',
+            selected: 'rgba(0, 0, 0, 1)',
+            disabledBackground: 'rgba(245, 245, 245, 1)',
+            focus: 'rgba(245, 245, 245, 1)',
+            disabled: 'rgba(196, 196, 196, 1)'
+        },
+        error: {
+            main: 'rgba(239, 75, 91, 1)',
+            dark: 'rgba(235, 30, 50, 1)',
+            light: 'rgba(243, 120, 132, 1)',
+            hover: 'rgba(253, 233, 235, 1)',
+            selected: 'rgba(251, 210, 214, 1)',
+            focusVisible: 'rgba(243, 120, 132, 1)',
+            outlinedBorder: 'rgba(235, 30, 50, 1)',
+            contrast: 'rgba(255, 255, 255, 1)'
+        },
+        warning: {
+            main: 'rgba(255, 219, 75, 1)',
+            dark: 'rgba(255, 210, 30, 1)',
+            light: 'rgba(255, 228, 120, 1)',
+            hover: 'rgba(255, 251, 233, 1)',
+            selected: 'rgba(255, 237, 165, 1)',
+            contrast: 'rgba(38, 38, 38, 1)',
+            focusVisible: 'rgba(255, 228, 120, 1)',
+            outlinedBorder: 'rgba(255, 210, 30, 1)'
+        },
+        info: {
+            main: 'rgba(0, 180, 240, 1)',
+            dark: 'rgba(0, 167, 238, 1)',
+            light: 'rgba(68, 193, 241, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(176, 228, 249, 1)',
+            focusVisible: 'rgba(224, 245, 253, 1)',
+            outlinedBorder: 'rgba(0, 167, 238, 1)',
+            contrast: 'rgba(38, 38, 38, 1)'
+        },
+        success: {
+            main: 'rgba(51, 179, 99, 1)',
+            dark: 'rgba(0, 160, 60, 1)',
+            light: 'rgba(102, 198, 138, 1)',
+            contrast: 'rgba(38, 38, 38, 1)',
+            hover: 'rgba(230, 246, 236, 1)',
+            selected: 'rgba(204, 236, 216, 1)',
+            focusVisible: 'rgba(102, 198, 138, 1)',
+            outlinedBorder: 'rgba(0, 160, 60, 1)'
+        },
+        common: {
+            white: 'rgba(255, 255, 255, 1)',
+            black: 'rgba(0, 0, 0, 1)',
+            grey: 'rgba(67, 67, 67, 1)'
+        },
+        background: {
+            'default': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-0': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-1': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-2': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-3': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-4': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-5': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-6': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-7': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-8': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-9': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-10': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-11': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-12': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-13': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-14': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-15': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-16': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-17': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-18': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-19': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-20': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-21': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-22': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-23': 'rgba(38, 38, 38, 1)',
+            'paper-elevation-24': 'rgba(38, 38, 38, 1)'
+        },
+        components: {
+            rating: {
+                enabledBorder: 'rgba(233, 233, 233, 1)',
+                activeFill: 'rgba(255, 210, 30, 1)'
+            },
+            avatar: {
+                fill: 'rgba(85, 85, 85, 1)'
+            },
+            input: {
+                standard: {
+                    enabledBorder: 'rgba(245, 245, 245, 1)',
+                    hoverBorder: 'rgba(255, 255, 255, 1)'
+                },
+                filled: {
+                    enabledFill: 'rgba(245, 245, 245, 1)',
+                    hoverFill: 'rgba(245, 245, 245, 1)'
+                },
+                outlined: {
+                    enabledBorder: 'rgba(217, 217, 217, 1)',
+                    hoverBorder: 'rgba(255, 255, 255, 1)'
+                }
+            },
+            switch: {
+                knobFillEnabled: 'rgba(196, 196, 196, 1)',
+                slideFill: 'rgba(245, 245, 245, 1)',
+                knowFillDisabled: 'rgba(85, 85, 85, 1)'
+            },
+            chip: {
+                defaultCloseFill: 'rgba(255, 255, 255, 1)',
+                defaultHoverFill: 'rgba(245, 245, 245, 1)',
+                defaultEnabledBorder: 'rgba(67, 67, 67, 1)',
+                defaultFocusFill: 'rgba(233, 233, 233, 1)'
+            },
+            tooltip: {
+                fill: 'rgba(123, 123, 123, 1)'
+            },
+            backdrop: {
+                fill: 'rgba(0, 0, 0, 0.5)'
+            },
+            appBar: {
+                defaultFill: 'rgba(38, 38, 38, 1)'
+            },
+            breadcrumbs: {
+                collapseFill: 'rgba(85, 85, 85, 1)'
+            },
+            alert: {
+                error: {
+                    color: 'rgba(253, 233, 235, 1)',
+                    background: 'rgba(229, 57, 53, 1)'
+                },
+                warning: {
+                    color: 'rgba(255, 246, 210, 1)',
+                    background: 'rgba(38, 38, 38, 1)'
+                },
+                info: {
+                    color: 'rgba(224, 245, 253, 1)',
+                    background: 'rgba(38, 38, 38, 1)'
+                },
+                success: {
+                    color: 'rgba(230, 246, 236, 1)',
+                    background: 'rgba(38, 38, 38, 1)'
+                }
+            },
+            stepper: {
+                connector: 'rgba(85, 85, 85, 1)'
+            },
+            snackbar: {
+                fill: 'rgba(38, 38, 38, 1)'
+            },
+            mediaImage: {
+                actions: 'rgba(196, 196, 196, 1)'
+            },
+            datagrid: {
+                border: 'rgba(123, 123, 123, 1)'
+            },
+            skeleton: {
+                border: 'rgba(123, 123, 123, 1)'
+            },
+            detail: {
+                close: {
+                    background: 'rgba(0, 0, 0, 1)',
+                    hover: 'rgba(38, 38, 38, 1)'
+                }
+            }
+        },
+        elevation: {
+            outlined: 'rgba(123, 123, 123, 1)',
+            Color: 'rgba(255, 255, 255, 1)'
+        },
+        _native: {
+            'scrollbar-bg': 'rgba(85, 85, 85, 1)'
+        },
+        contrast: {
+            contrastText: 'rgba(0, 167, 238, 1)',
+            main: 'rgba(255, 255, 255, 1)',
+            dark: 'rgba(255, 255, 255, 1)',
+            light: 'rgba(255, 255, 255, 1)',
+            contrast: 'rgba(255, 255, 255, 1)',
+            hover: 'rgba(224, 245, 253, 1)',
+            selected: 'rgba(224, 245, 253, 1)',
+            focus: 'rgba(176, 228, 249, 1)',
+            focusVisible: 'rgba(224, 245, 253, 1)',
+            outlinedBorder: 'rgba(224, 245, 253, 1)'
+        }
+    }
+};