+++ src/theme/index.ts
@@ -1,0 +1,6 @@
+export * from './theme';
+export * from './config/breakpoints';
+export * from './config/palette';
+export * from './config/shape';
+export * from './config/spacing';
+export * from './config/typography';