+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.0] - 2023-10-30
+### Added
+- [@jpalis]: Added config files for theme creation (breakpoints, palette, shape, spacing, typography).
+- [@jpalis]: Added getTheme() export with mui theme.
+- [@jpalis]: Added `ThemeConfig` component with MUI `ThemeProvider` inside. 
+- [@jpalis]: Added test for getTheme() and `ThemeConfig`.
+
 ## [0.0.0] - 2023-10-21
 ### Added
 - [@jpalis]: Initial commit.